+++ server/packages/super_simple_authentication_toolkit/lib/src/integrations/3rd_party_providers/3rd_party_providers.dart
@@ -1,0 +1,5 @@
+// Ignored to allow for number in filename
+// ignore_for_file: file_names
+
+export 'sign_in_with_apple.dart';
+export 'sign_in_with_google.dart';