+++ server/packages/super_simple_authentication_toolkit/lib/src/integrations/email/email_provider.dart
@@ -1,0 +1,21 @@
+/// {@template email_provider}
+/// An abstract interface for email providers.
+/// {@endtemplate}
+abstract class EmailProvider {
+  /// Sends an email to the given recipient with the given subject and body.
+  Future<void> sendEmail({
+    required String to,
+    required String subject,
+    required String body,
+    required String from,
+  });
+
+  /// Sends an email using a Sendgrid dynamic template.
+  Future<void> sendEmailWithTemplate({
+    required String to,
+    required String templateId,
+    required Map<String, dynamic> dynamicTemplateData,
+    required String from,
+    String? subject,
+  });
+}