+++ server/packages/super_simple_authentication_toolkit/lib/src/integrations/sms/fake_sms.dart
@@ -1,0 +1,23 @@
+import 'dart:io';
+
+import 'package:super_simple_authentication_toolkit/super_simple_authentication_toolkit.dart';
+
+/// {@template fake_sms}
+/// A fake SMS service for testing.
+/// {@endtemplate}
+class FakeSms implements SmsProvider {
+  /// {@macro fake_sms}
+  const FakeSms({required Stdout stdout}) : _stdout = stdout;
+
+  final Stdout _stdout;
+
+  @override
+  Future<void> sendSms(String to, String body) async {
+    _stdout.writeln(
+      '''
+Sending SMS
+[TO] $to
+$body''',
+    );
+  }
+}